+++ package.json
@@ -19,7 +19,6 @@
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "express-async-handler": "^1.2.0",
-    "express-session": "^1.17.2",
     "express-validator": "^6.14.0",
     "google-auth-library": "^7.14.1",
     "helmet": "^5.0.2",
@@ -31,7 +30,6 @@
     "rate-limiter-flexible": "^2.3.6",
     "redis": "^4.0.6",
     "swagger-autogen": "^2.20.8",
-    "swagger-jsdoc": "^6.2.0",
     "swagger-ui-express": "^4.3.0",
     "winston": "^3.7.2"
   },
