+++ src/utils/generateToken.js
@@ -1,9 +1,7 @@
 const jwt = require('jsonwebtoken')
 
-const generateToken = (id) => {
+exports.generateToken = (id) => {
     return jwt.sign({ id }, process.env.JWT_SECRET, {
         expiresIn: `${process.env.JWT_EXPIRE}`,
     })
 }
-
-module.exports = generateToken
